+++ src/nypl_py_utils/functions/log_helper.py
@@ -3,11 +3,11 @@ import os
 import sys
 
 levels = {
-    "debug": logging.DEBUG,
-    "info": logging.INFO,
-    "warning": logging.WARNING,
-    "error": logging.ERROR,
-    "critical": logging.CRITICAL,
+    'debug': logging.DEBUG,
+    'info': logging.INFO,
+    'warning': logging.WARNING,
+    'error': logging.ERROR,
+    'critical': logging.CRITICAL
 }
 
 
@@ -18,12 +18,13 @@ def create_log(module):
 
     console_log = logging.StreamHandler(stream=sys.stdout)
 
-    log_level = os.environ.get("LOG_LEVEL", "info").lower()
+    log_level = os.environ.get('LOG_LEVEL', 'info').lower()
 
     logger.setLevel(levels[log_level])
     console_log.setLevel(levels[log_level])
 
-    formatter = logging.Formatter("%(asctime)s | %(name)s | %(levelname)s: %(message)s")
+    formatter = logging.Formatter(
+        '%(asctime)s | %(name)s | %(levelname)s: %(message)s')
     console_log.setFormatter(formatter)
 
     logger.addHandler(console_log)